7.1
Write a code which take source file and destination file as input a command line argument.  It copies the source file contents to destination file. If source file does not exist, it should give appropiate message to the user.  If destination file does not exist, it should be created. If it exists, program should ask that, "whether you want to overwrite?(yes/no)".